+++ pkg/kubelet/dockertools/docker.go
@@ -18,24 +18,17 @@ package dockertools
 
 import (
 	"fmt"
-	"math/rand"
 	"net/http"
 	"path"
-	"strconv"
 	"strings"
-	"time"
 
-	dockerdigest "github.com/docker/distribution/digest"
-	dockerref "github.com/docker/distribution/reference"
 	"github.com/docker/docker/pkg/jsonmessage"
-	dockerapi "github.com/docker/engine-api/client"
 	dockertypes "github.com/docker/engine-api/types"
 	"github.com/golang/glog"
-	"k8s.io/apimachinery/pkg/types"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/credentialprovider"
-	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
+	"k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
 	"k8s.io/kubernetes/pkg/kubelet/images"
 )
 
@@ -44,60 +37,11 @@ const (
 	ext4MaxFileNameLen = 255
 
 	DockerType = "docker"
-
-	// https://docs.docker.com/engine/reference/api/docker_remote_api/
-	// docker version should be at least 1.10.x
-	minimumDockerAPIVersion = "1.22"
-
-	statusRunningPrefix = "Up"
-	statusExitedPrefix  = "Exited"
-	statusCreatedPrefix = "Created"
 )
 
-// DockerInterface is an abstract interface for testability.  It abstracts the interface of docker client.
-type DockerInterface interface {
-	ListContainers(options dockertypes.ContainerListOptions) ([]dockertypes.Container, error)
-	InspectContainer(id string) (*dockertypes.ContainerJSON, error)
-	CreateContainer(dockertypes.ContainerCreateConfig) (*dockertypes.ContainerCreateResponse, error)
-	StartContainer(id string) error
-	StopContainer(id string, timeout int) error
-	RemoveContainer(id string, opts dockertypes.ContainerRemoveOptions) error
-	InspectImageByRef(imageRef string) (*dockertypes.ImageInspect, error)
-	InspectImageByID(imageID string) (*dockertypes.ImageInspect, error)
-	ListImages(opts dockertypes.ImageListOptions) ([]dockertypes.Image, error)
-	PullImage(image string, auth dockertypes.AuthConfig, opts dockertypes.ImagePullOptions) error
-	RemoveImage(image string, opts dockertypes.ImageRemoveOptions) ([]dockertypes.ImageDelete, error)
-	ImageHistory(id string) ([]dockertypes.ImageHistory, error)
-	Logs(string, dockertypes.ContainerLogsOptions, StreamOptions) error
-	Version() (*dockertypes.Version, error)
-	Info() (*dockertypes.Info, error)
-	CreateExec(string, dockertypes.ExecConfig) (*dockertypes.ContainerExecCreateResponse, error)
-	StartExec(string, dockertypes.ExecStartCheck, StreamOptions) error
-	InspectExec(id string) (*dockertypes.ContainerExecInspect, error)
-	AttachToContainer(string, dockertypes.ContainerAttachOptions, StreamOptions) error
-	ResizeContainerTTY(id string, height, width int) error
-	ResizeExecTTY(id string, height, width int) error
-}
-
-// KubeletContainerName encapsulates a pod name and a Kubernetes container name.
-type KubeletContainerName struct {
-	PodFullName   string
-	PodUID        types.UID
-	ContainerName string
-}
-
-// containerNamePrefix is used to identify the containers on the node managed by this
-// process.
-var containerNamePrefix = "k8s"
-
-// SetContainerNamePrefix allows the container prefix name for this process to be changed.
-// This is intended to support testing and bootstrapping experimentation. It cannot be
-// changed once the Kubelet starts.
-func SetContainerNamePrefix(prefix string) {
-	containerNamePrefix = prefix
-}
-
 // DockerPuller is an abstract interface for testability.  It abstracts image pull operations.
+// DockerPuller is *not* in use anywhere in the codebase.
+// TODO: Examine whether we can migrate the unit tests and remove the code.
 type DockerPuller interface {
 	Pull(image string, secrets []v1.Secret) error
 	GetImageRef(image string) (string, error)
@@ -105,12 +49,12 @@ type DockerPuller interface {
 
 // dockerPuller is the default implementation of DockerPuller.
 type dockerPuller struct {
-	client  DockerInterface
+	client  libdocker.Interface
 	keyring credentialprovider.DockerKeyring
 }
 
 // newDockerPuller creates a new instance of the default implementation of DockerPuller.
-func newDockerPuller(client DockerInterface) DockerPuller {
+func newDockerPuller(client libdocker.Interface) DockerPuller {
 	return &dockerPuller{
 		client:  client,
 		keyring: credentialprovider.NewDockerKeyring(),
@@ -131,104 +75,6 @@ func filterHTTPError(err error, image string) error {
 	} else {
 		return err
 	}
-}
-
-// matchImageTagOrSHA checks if the given image specifier is a valid image ref,
-// and that it matches the given image. It should fail on things like image IDs
-// (config digests) and other digest-only references, but succeed on image names
-// (`foo`), tag references (`foo:bar`), and manifest digest references
-// (`foo@sha256:xyz`).
-func matchImageTagOrSHA(inspected dockertypes.ImageInspect, image string) bool {
-	// The image string follows the grammar specified here
-	// https://github.com/docker/distribution/blob/master/reference/reference.go#L4
-	named, err := dockerref.ParseNamed(image)
-	if err != nil {
-		glog.V(4).Infof("couldn't parse image reference %q: %v", image, err)
-		return false
-	}
-	_, isTagged := named.(dockerref.Tagged)
-	digest, isDigested := named.(dockerref.Digested)
-	if !isTagged && !isDigested {
-		// No Tag or SHA specified, so just return what we have
-		return true
-	}
-
-	if isTagged {
-		// Check the RepoTags for a match.
-		for _, tag := range inspected.RepoTags {
-			// An image name (without the tag/digest) can be [hostname '/'] component ['/' component]*
-			// Because either the RepoTag or the name *may* contain the
-			// hostname or not, we only check for the suffix match.
-			if strings.HasSuffix(image, tag) || strings.HasSuffix(tag, image) {
-				return true
-			}
-		}
-	}
-
-	if isDigested {
-		for _, repoDigest := range inspected.RepoDigests {
-			named, err := dockerref.ParseNamed(repoDigest)
-			if err != nil {
-				glog.V(4).Infof("couldn't parse image RepoDigest reference %q: %v", repoDigest, err)
-				continue
-			}
-			if d, isDigested := named.(dockerref.Digested); isDigested {
-				if digest.Digest().Algorithm().String() == d.Digest().Algorithm().String() &&
-					digest.Digest().Hex() == d.Digest().Hex() {
-					return true
-				}
-			}
-		}
-
-		// process the ID as a digest
-		id, err := dockerdigest.ParseDigest(inspected.ID)
-		if err != nil {
-			glog.V(4).Infof("couldn't parse image ID reference %q: %v", id, err)
-			return false
-		}
-		if digest.Digest().Algorithm().String() == id.Algorithm().String() && digest.Digest().Hex() == id.Hex() {
-			return true
-		}
-	}
-	glog.V(4).Infof("Inspected image (%q) does not match %s", inspected.ID, image)
-	return false
-}
-
-// matchImageIDOnly checks that the given image specifier is a digest-only
-// reference, and that it matches the given image.
-func matchImageIDOnly(inspected dockertypes.ImageInspect, image string) bool {
-	// If the image ref is literally equal to the inspected image's ID,
-	// just return true here (this might be the case for Docker 1.9,
-	// where we won't have a digest for the ID)
-	if inspected.ID == image {
-		return true
-	}
-
-	// Otherwise, we should try actual parsing to be more correct
-	ref, err := dockerref.Parse(image)
-	if err != nil {
-		glog.V(4).Infof("couldn't parse image reference %q: %v", image, err)
-		return false
-	}
-
-	digest, isDigested := ref.(dockerref.Digested)
-	if !isDigested {
-		glog.V(4).Infof("the image reference %q was not a digest reference")
-		return false
-	}
-
-	id, err := dockerdigest.ParseDigest(inspected.ID)
-	if err != nil {
-		glog.V(4).Infof("couldn't parse image ID reference %q: %v", id, err)
-		return false
-	}
-
-	if digest.Digest().Algorithm().String() == id.Algorithm().String() && digest.Digest().Hex() == id.Hex() {
-		return true
-	}
-
-	glog.V(4).Infof("The reference %s does not directly refer to the given image's ID (%q)", image, inspected.ID)
-	return false
 }
 
 func (p dockerPuller) Pull(image string, secrets []v1.Secret) error {
@@ -296,61 +142,10 @@ func (p dockerPuller) GetImageRef(image string) (string, error) {
 		}
 		return imageRef, nil
 	}
-	if IsImageNotFoundError(err) {
+	if libdocker.IsImageNotFoundError(err) {
 		return "", nil
 	}
 	return "", err
-}
-
-// Creates a name which can be reversed to identify both full pod name and container name.
-// This function returns stable name, unique name and a unique id.
-// Although rand.Uint32() is not really unique, but it's enough for us because error will
-// only occur when instances of the same container in the same pod have the same UID. The
-// chance is really slim.
-func BuildDockerName(dockerName KubeletContainerName, container *v1.Container) (string, string, string) {
-	containerName := dockerName.ContainerName + "." + strconv.FormatUint(kubecontainer.HashContainerLegacy(container), 16)
-	stableName := fmt.Sprintf("%s_%s_%s_%s",
-		containerNamePrefix,
-		containerName,
-		dockerName.PodFullName,
-		dockerName.PodUID)
-	UID := fmt.Sprintf("%08x", rand.Uint32())
-	return stableName, fmt.Sprintf("%s_%s", stableName, UID), UID
-}
-
-// Unpacks a container name, returning the pod full name and container name we would have used to
-// construct the docker name. If we are unable to parse the name, an error is returned.
-func ParseDockerName(name string) (dockerName *KubeletContainerName, hash uint64, err error) {
-	// For some reason docker appears to be appending '/' to names.
-	// If it's there, strip it.
-	name = strings.TrimPrefix(name, "/")
-	parts := strings.Split(name, "_")
-	if len(parts) == 0 || parts[0] != containerNamePrefix {
-		err = fmt.Errorf("failed to parse Docker container name %q into parts", name)
-		return nil, 0, err
-	}
-	if len(parts) < 6 {
-		// We have at least 5 fields.  We may have more in the future.
-		// Anything with less fields than this is not something we can
-		// manage.
-		glog.Warningf("found a container with the %q prefix, but too few fields (%d): %q", containerNamePrefix, len(parts), name)
-		err = fmt.Errorf("Docker container name %q has less parts than expected %v", name, parts)
-		return nil, 0, err
-	}
-
-	nameParts := strings.Split(parts[1], ".")
-	containerName := nameParts[0]
-	if len(nameParts) > 1 {
-		hash, err = strconv.ParseUint(nameParts[1], 16, 32)
-		if err != nil {
-			glog.Warningf("invalid container hash %q in container %q", nameParts[1], name)
-		}
-	}
-
-	podFullName := parts[2] + "_" + parts[3]
-	podUID := types.UID(parts[4])
-
-	return &KubeletContainerName{podFullName, podUID, containerName}, hash, nil
 }
 
 func LogSymlink(containerLogsDir, podFullName, containerName, dockerId string) string {
@@ -361,58 +156,4 @@ func LogSymlink(containerLogsDir, podFullName, containerName, dockerId string) s
 		logPath = logPath[:ext4MaxFileNameLen-len(suffix)]
 	}
 	return path.Join(containerLogsDir, logPath+suffix)
-}
-
-// Get a *dockerapi.Client, either using the endpoint passed in, or using
-// DOCKER_HOST, DOCKER_TLS_VERIFY, and DOCKER_CERT path per their spec
-func getDockerClient(dockerEndpoint string) (*dockerapi.Client, error) {
-	if len(dockerEndpoint) > 0 {
-		glog.Infof("Connecting to docker on %s", dockerEndpoint)
-		return dockerapi.NewClient(dockerEndpoint, "", nil, nil)
-	}
-	return dockerapi.NewEnvClient()
-}
-
-// ConnectToDockerOrDie creates docker client connecting to docker daemon.
-// If the endpoint passed in is "fake://", a fake docker client
-// will be returned. The program exits if error occurs. The requestTimeout
-// is the timeout for docker requests. If timeout is exceeded, the request
-// will be cancelled and throw out an error. If requestTimeout is 0, a default
-// value will be applied.
-func ConnectToDockerOrDie(dockerEndpoint string, requestTimeout, imagePullProgressDeadline time.Duration) DockerInterface {
-	if dockerEndpoint == "fake://" {
-		return NewFakeDockerClient()
-	}
-	client, err := getDockerClient(dockerEndpoint)
-	if err != nil {
-		glog.Fatalf("Couldn't connect to docker: %v", err)
-	}
-	glog.Infof("Start docker client with request timeout=%v", requestTimeout)
-	return newKubeDockerClient(client, requestTimeout, imagePullProgressDeadline)
-}
-
-// GetKubeletDockerContainers lists all container or just the running ones.
-// Returns a list of docker containers that we manage
-// TODO: This function should be deleted after migrating
-// test/e2e_node/garbage_collector_test.go off of it.
-func GetKubeletDockerContainers(client DockerInterface, allContainers bool) ([]*dockertypes.Container, error) {
-	result := []*dockertypes.Container{}
-	containers, err := client.ListContainers(dockertypes.ContainerListOptions{All: allContainers})
-	if err != nil {
-		return nil, err
-	}
-	for i := range containers {
-		container := &containers[i]
-		if len(container.Names) == 0 {
-			continue
-		}
-		// Skip containers that we didn't create to allow users to manually
-		// spin up their own containers if they want.
-		if !strings.HasPrefix(container.Names[0], "/"+containerNamePrefix+"_") {
-			glog.V(5).Infof("Docker Container: %s is not managed by kubelet.", container.Names[0])
-			continue
-		}
-		result = append(result, container)
-	}
-	return result, nil
 }
