+++ pkg/kubelet/dockertools/docker_test.go
@@ -19,371 +19,20 @@ package dockertools
 import (
 	"encoding/json"
 	"fmt"
-	"hash/adler32"
 	"math/rand"
 	"path"
-	"reflect"
 	"strings"
 	"testing"
 
 	"github.com/docker/docker/pkg/jsonmessage"
-	dockertypes "github.com/docker/engine-api/types"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/credentialprovider"
+	"k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
 	"k8s.io/kubernetes/pkg/kubelet/images"
-	hashutil "k8s.io/kubernetes/pkg/util/hash"
 )
 
-func verifyCalls(t *testing.T, fakeDocker *FakeDockerClient, calls []string) {
-	assert.New(t).NoError(fakeDocker.AssertCalls(calls))
-}
-
-func verifyStringArrayEquals(t *testing.T, actual, expected []string) {
-	invalid := len(actual) != len(expected)
-	if !invalid {
-		for ix, value := range actual {
-			if expected[ix] != value {
-				invalid = true
-			}
-		}
-	}
-	if invalid {
-		t.Errorf("Expected: %#v, Actual: %#v", expected, actual)
-	}
-}
-
-func findPodContainer(dockerContainers []*dockertypes.Container, podFullName string, uid types.UID, containerName string) (*dockertypes.Container, bool, uint64) {
-	for _, dockerContainer := range dockerContainers {
-		if len(dockerContainer.Names) == 0 {
-			continue
-		}
-		dockerName, hash, err := ParseDockerName(dockerContainer.Names[0])
-		if err != nil {
-			continue
-		}
-		if dockerName.PodFullName == podFullName &&
-			(uid == "" || dockerName.PodUID == uid) &&
-			dockerName.ContainerName == containerName {
-			return dockerContainer, true, hash
-		}
-	}
-	return nil, false, 0
-}
-
-func TestGetContainerID(t *testing.T) {
-	fakeDocker := NewFakeDockerClient()
-	fakeDocker.SetFakeRunningContainers([]*FakeContainer{
-		{
-			ID:   "foobar",
-			Name: "/k8s_foo_qux_ns_1234_42",
-		},
-		{
-			ID:   "barbar",
-			Name: "/k8s_bar_qux_ns_2565_42",
-		},
-	})
-
-	dockerContainers, err := GetKubeletDockerContainers(fakeDocker, false)
-	if err != nil {
-		t.Errorf("Expected no error, Got %#v", err)
-	}
-	if len(dockerContainers) != 2 {
-		t.Errorf("Expected %#v, Got %#v", fakeDocker.RunningContainerList, dockerContainers)
-	}
-	verifyCalls(t, fakeDocker, []string{"list"})
-
-	dockerContainer, found, _ := findPodContainer(dockerContainers, "qux_ns", "", "foo")
-	if dockerContainer == nil || !found {
-		t.Errorf("Failed to find container %#v", dockerContainer)
-	}
-
-	fakeDocker.ClearCalls()
-	dockerContainer, found, _ = findPodContainer(dockerContainers, "foobar", "", "foo")
-	verifyCalls(t, fakeDocker, []string{})
-	if dockerContainer != nil || found {
-		t.Errorf("Should not have found container %#v", dockerContainer)
-	}
-}
-
-func verifyPackUnpack(t *testing.T, podNamespace, podUID, podName, containerName string) {
-	container := &v1.Container{Name: containerName}
-	hasher := adler32.New()
-	hashutil.DeepHashObject(hasher, *container)
-	computedHash := uint64(hasher.Sum32())
-	podFullName := fmt.Sprintf("%s_%s", podName, podNamespace)
-	_, name, _ := BuildDockerName(KubeletContainerName{podFullName, types.UID(podUID), container.Name}, container)
-	returned, hash, err := ParseDockerName(name)
-	if err != nil {
-		t.Errorf("Failed to parse Docker container name %q: %v", name, err)
-	}
-	if podFullName != returned.PodFullName || podUID != string(returned.PodUID) || containerName != returned.ContainerName || computedHash != hash {
-		t.Errorf("For (%s, %s, %s, %d), unpacked (%s, %s, %s, %d)", podFullName, podUID, containerName, computedHash, returned.PodFullName, returned.PodUID, returned.ContainerName, hash)
-	}
-}
-
-func TestContainerNaming(t *testing.T) {
-	podUID := "12345678"
-	verifyPackUnpack(t, "file", podUID, "name", "container")
-	verifyPackUnpack(t, "file", podUID, "name-with-dashes", "container")
-	// UID is same as pod name
-	verifyPackUnpack(t, "file", podUID, podUID, "container")
-	// No Container name
-	verifyPackUnpack(t, "other", podUID, "name", "")
-
-	container := &v1.Container{Name: "container"}
-	podName := "foo"
-	podNamespace := "test"
-	name := fmt.Sprintf("k8s_%s_%s_%s_%s_42", container.Name, podName, podNamespace, podUID)
-	podFullName := fmt.Sprintf("%s_%s", podName, podNamespace)
-
-	returned, hash, err := ParseDockerName(name)
-	if err != nil {
-		t.Errorf("Failed to parse Docker container name %q: %v", name, err)
-	}
-	if returned.PodFullName != podFullName || string(returned.PodUID) != podUID || returned.ContainerName != container.Name || hash != 0 {
-		t.Errorf("unexpected parse: %s %s %s %d", returned.PodFullName, returned.PodUID, returned.ContainerName, hash)
-	}
-}
-
-func TestMatchImageTagOrSHA(t *testing.T) {
-	for i, testCase := range []struct {
-		Inspected dockertypes.ImageInspect
-		Image     string
-		Output    bool
-	}{
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"ubuntu:latest"}},
-			Image:     "ubuntu",
-			Output:    true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"ubuntu:14.04"}},
-			Image:     "ubuntu:latest",
-			Output:    false,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"colemickens/hyperkube-amd64:217.9beff63"}},
-			Image:     "colemickens/hyperkube-amd64:217.9beff63",
-			Output:    true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"colemickens/hyperkube-amd64:217.9beff63"}},
-			Image:     "docker.io/colemickens/hyperkube-amd64:217.9beff63",
-			Output:    true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"docker.io/kubernetes/pause:latest"}},
-			Image:     "kubernetes/pause:latest",
-			Output:    true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			Output: true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208f7a29005",
-			Output: false,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208",
-			Output: false,
-		},
-		{
-			// mismatched ID is ignored
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:0000f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			Output: false,
-		},
-		{
-			// invalid digest is ignored
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:unparseable",
-			},
-			Image:  "myimage@sha256:unparseable",
-			Output: false,
-		},
-		{
-			// v1 schema images can be pulled in one format and returned in another
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: true,
-		},
-		{
-			// RepoDigest match is is required
-			Inspected: dockertypes.ImageInspect{
-				ID:          "",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:000084acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: false,
-		},
-		{
-			// RepoDigest match is allowed
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: true,
-		},
-		{
-			// RepoDigest and ID are checked
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: true,
-		},
-		{
-			// unparseable RepoDigests are skipped
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{
-					"centos/ruby-23-centos7@sha256:unparseable",
-					"docker.io/centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-				},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: true,
-		},
-		{
-			// unparseable RepoDigest is ignored
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:unparseable"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: false,
-		},
-		{
-			// unparseable image digest is ignored
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:unparseable"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:unparseable",
-			Output: false,
-		},
-		{
-			// prefix match is rejected for ID and RepoDigest
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:unparseable",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:unparseable"},
-			},
-			Image:  "sha256:unparseable",
-			Output: false,
-		},
-		{
-			// possible SHA prefix match is rejected for ID and RepoDigest because it is not in the named format
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:0000f247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"docker.io/centos/ruby-23-centos7@sha256:0000f247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227"},
-			},
-			Image:  "sha256:0000",
-			Output: false,
-		},
-	} {
-		match := matchImageTagOrSHA(testCase.Inspected, testCase.Image)
-		assert.Equal(t, testCase.Output, match, testCase.Image+fmt.Sprintf(" is not a match (%d)", i))
-	}
-}
-
-func TestMatchImageIDOnly(t *testing.T) {
-	for i, testCase := range []struct {
-		Inspected dockertypes.ImageInspect
-		Image     string
-		Output    bool
-	}{
-		// shouldn't match names or tagged names
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"ubuntu:latest"}},
-			Image:     "ubuntu",
-			Output:    false,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{RepoTags: []string{"colemickens/hyperkube-amd64:217.9beff63"}},
-			Image:     "colemickens/hyperkube-amd64:217.9beff63",
-			Output:    false,
-		},
-		// should match name@digest refs if they refer to the image ID (but only the full ID)
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			Output: true,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208f7a29005",
-			Output: false,
-		},
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:2208",
-			Output: false,
-		},
-		// should match when the IDs are literally the same
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "foobar",
-			},
-			Image:  "foobar",
-			Output: true,
-		},
-		// shouldn't match mismatched IDs
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:2208f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			},
-			Image:  "myimage@sha256:0000f7a29005d226d1ee33a63e33af1f47af6156c740d7d23c7948e8d282d53d",
-			Output: false,
-		},
-		// shouldn't match invalid IDs or refs
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID: "sha256:unparseable",
-			},
-			Image:  "myimage@sha256:unparseable",
-			Output: false,
-		},
-		// shouldn't match against repo digests
-		{
-			Inspected: dockertypes.ImageInspect{
-				ID:          "sha256:9bbdf247c91345f0789c10f50a57e36a667af1189687ad1de88a6243d05a2227",
-				RepoDigests: []string{"centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf"},
-			},
-			Image:  "centos/ruby-23-centos7@sha256:940584acbbfb0347272112d2eb95574625c0c60b4e2fdadb139de5859cf754bf",
-			Output: false,
-		},
-	} {
-		match := matchImageIDOnly(testCase.Inspected, testCase.Image)
-		assert.Equal(t, testCase.Output, match, fmt.Sprintf("%s is not a match (%d)", testCase.Image, i))
-	}
-
-}
-
+// TODO: Examine the tests and see if they can be migrated to kuberuntime.
 func TestPullWithNoSecrets(t *testing.T) {
 	tests := []struct {
 		imageName     string
@@ -399,7 +48,7 @@ func TestPullWithNoSecrets(t *testing.T) {
 	}
 	for _, test := range tests {
 		fakeKeyring := &credentialprovider.FakeKeyring{}
-		fakeClient := NewFakeDockerClient()
+		fakeClient := libdocker.NewFakeDockerClient()
 
 		dp := dockerPuller{
 			client:  fakeClient,
@@ -412,13 +61,8 @@ func TestPullWithNoSecrets(t *testing.T) {
 			continue
 		}
 
-		if e, a := 1, len(fakeClient.pulled); e != a {
-			t.Errorf("%s: expected 1 pulled image, got %d: %v", test.imageName, a, fakeClient.pulled)
-			continue
-		}
-
-		if e, a := test.expectedImage, fakeClient.pulled[0]; e != a {
-			t.Errorf("%s: expected pull of %q, but got %q", test.imageName, e, a)
+		if err := fakeClient.AssertImagesPulled([]string{test.imageName}); err != nil {
+			t.Errorf("images pulled do not match the expected: %v", err)
 		}
 	}
 }
@@ -442,7 +86,7 @@ func TestPullWithJSONError(t *testing.T) {
 	}
 	for i, test := range tests {
 		fakeKeyring := &credentialprovider.FakeKeyring{}
-		fakeClient := NewFakeDockerClient()
+		fakeClient := libdocker.NewFakeDockerClient()
 		fakeClient.InjectError("pull", test.err)
 
 		puller := &dockerPuller{
@@ -520,7 +164,7 @@ func TestPullWithSecrets(t *testing.T) {
 		builtInKeyRing := &credentialprovider.BasicDockerKeyring{}
 		builtInKeyRing.Add(test.builtInDockerConfig)
 
-		fakeClient := NewFakeDockerClient()
+		fakeClient := libdocker.NewFakeDockerClient()
 
 		dp := dockerPuller{
 			client:  fakeClient,
@@ -532,14 +176,8 @@ func TestPullWithSecrets(t *testing.T) {
 			t.Errorf("%s: unexpected non-nil err: %s", i, err)
 			continue
 		}
-
-		if e, a := 1, len(fakeClient.pulled); e != a {
-			t.Errorf("%s: expected 1 pulled image, got %d: %v", i, a, fakeClient.pulled)
-			continue
-		}
-
-		if e, a := test.expectedPulls, fakeClient.pulled; !reflect.DeepEqual(e, a) {
-			t.Errorf("%s: expected pull of %v, but got %v", i, e, a)
+		if err := fakeClient.AssertImagesPulledMsgs(test.expectedPulls); err != nil {
+			t.Errorf("images pulled do not match the expected: %v", err)
 		}
 	}
 }
